+++ pystiche_papers/utils/__init__.py
@@ -1,3 +1,4 @@
 from .cuda import *
 from .misc import *
 from .modules import *
+from .transforms import *
